Hub Room/Collisions, Sprite Rendering, Multiple Cameras

+++ DX11 Game/Graphics/DepthStencil.h
@@ -15,7 +15,8 @@ namespace Bind
 		{
 			try
 			{
-				CD3D11_TEXTURE2D_DESC depthStencilDesc( DXGI_FORMAT_D24_UNORM_S8_UINT, width, height );
+				CD3D11_TEXTURE2D_DESC depthStencilDesc( DXGI_FORMAT_D24_UNORM_S8_UINT,
+					static_cast<UINT>( width ), static_cast<UINT>( height ) );
 				depthStencilDesc.MipLevels = 1;
 				depthStencilDesc.SampleDesc.Count = 1;
 				depthStencilDesc.SampleDesc.Quality = 0;
